Problem Statement
Davy Jones needs to conquer a ship for his pirates. But the only rule he has to follow is to conquer only those ships whose names are partially mirrored at the end.


For Example: If a ship's name is 'abxyzba' then he can conquer the ship because the string has mirrored string at the beginning and at the end i.e 'ab'.



Now your job is to develop a programming code which helps Jones to identify the ships.