Problem Statement
Tris (or as her parents call her Beatrice) is going through a series of training and evaluation to get into top 10 and remain in Dauntless Faction. She is currently placed 12th and will become faction-less if she fails, today.

Four (or Tobias) gives Tris a string, consisting of 'i'(s) and 'd'(s) only. Tris has to select a number and increment it by 1 when she encounters 'i' and reduce it by 1 when she encounters 'd'. Thus calculate the final result.

There are several results possible, but Four wants the starting number and the corresponding result, such that their sum is minimum. However there is a catch, like in all Dauntless exercises, the number cannot be negative at any step.