Problem Statement
Gate returns 1 if all its inputs are 1, otherwise 0.
Input:
The first line of input takes the number of test cases, T. Then T test cases follow.
Each test case consists of 2 lines.
The first line of each test case takes the number of inputs to the AND Gate, N. The second line of each test case takes N space separated integers denoting the inputs to the AND Gate. Note that the inputs can be either 1's or 0's.
Output:
Print the output of the N input AND Gate for each test case on a new line.